Extract logic to survey url provider.
