+++ src/xcode/ENA/ENA/Source/Scenes/ExposureDetection/__tests__/Survey/SurveyConsentViewModelTests.swift
@@ -8,17 +8,8 @@ import XCTest
 class SurveyConsentViewModelTests: XCTestCase {
 
 	func test_WHEN_dynamicTableViewModel_THEN_CorrectModelIsReturned() throws {
-		let store = MockTestStore()
-		let client = ClientMock()
-		let otpService = OTPService(store: store, client: client)
-
-		let deviceCheck = PPACDeviceCheckMock(true, deviceToken: "SomeToken")
-		let ppacService = try PPACService(store: store, deviceCheck: deviceCheck)
-
 		let viewModel = SurveyConsentViewModel(
-			configurationProvider: CachedAppConfigurationMock(),
-			ppacService: ppacService,
-			otpService: otpService
+			surveyURLProvider: SurveyURLProviderFake()
 		)
 
 		let dynamicTableViewModel = viewModel.dynamicTableViewModel
@@ -28,31 +19,8 @@ class SurveyConsentViewModelTests: XCTestCase {
 		XCTAssertEqual(dynamicTableViewModel.section(1).cells.count, 1)
 		XCTAssertEqual(dynamicTableViewModel.section(2).cells.count, 1)
 	}
+}
 
-	func test_WHEN_getURLIsCalled_THEN_AnURLIstReturned() throws {
-		let store = MockTestStore()
-		let client = ClientMock()
-		let otpService = OTPService(store: store, client: client)
-
-		let deviceCheck = PPACDeviceCheckMock(true, deviceToken: "SomeToken")
-		let ppacService = try PPACService(store: store, deviceCheck: deviceCheck)
-
-		let viewModel = SurveyConsentViewModel(
-			configurationProvider: CachedAppConfigurationMock(),
-			ppacService: ppacService,
-			otpService: otpService
-		)
-
-		let urlExpectation = expectation(description: "URL is returned.")
-		viewModel.getURL { result in
-			switch result {
-			case .success:
-				urlExpectation.fulfill()
-			case .failure:
-				XCTFail("Error not expected.")
-			}
-		}
-
-		waitForExpectations(timeout: .long)
-	}
+private struct SurveyURLProviderFake: SurveyURLProvidable {
+	func getURL(_ completion: @escaping (Result<URL, SurveyConsentError>) -> Void) { }
 }
